Struct indy_vdr::common::merkle_tree::MerkleTree

pub struct MerkleTree {
    pub root: Tree,
    pub height: usize,
    pub count: usize,
    pub nodes_count: usize,
}

A Merkle tree is a binary tree, with values of type T at the leafs, and where every internal node holds the hash of the concatenation of the hashes of its children nodes.

Fields
root: Tree

The root of the inner binary tree

height: usize

The height of the tree

count: usize

The number of leaf nodes in the tree

nodes_count: usize

The number of nodes in the tree

Methods

impl MerkleTree

pub fn from_vec(values: Vec<Vec<u8>>) -> Result<Self, ValidationError>

Constructs a Merkle Tree from a vector of data blocks. Returns None if values is empty.

pub fn root_hash(&self) -> &Vec<u8>

Returns the root hash of Merkle tree

pub fn root_hash_hex(&self) -> String

Returns the hex root hash of Merkle tree

pub fn height(&self) -> usize

Returns the height of Merkle tree

pub fn count(&self) -> usize

Returns the number of leaves in the Merkle tree

pub fn is_empty(&self) -> bool

Returns whether the Merkle tree is empty or not

pub fn gen_proof(
&self,
value: Vec<u8>
) -> Result<Option<Proof>, ValidationError>

Generate an inclusion proof for the given value. Returns None if the given value is not found in the tree.

pub fn iter(&self) -> LeavesIterator

Creates an Iterator over the values contained in this Merkle tree.

impl MerkleTree

pub fn find_hash<'a>(
from: &'a Tree,
required_hash: &Vec<u8>
) -> Option<&'a Tree>

pub fn consistency_proof(
&self,
new_root_hash: &Vec<u8>,
new_size: usize,
proof: &Vec<Vec<u8>>
) -> Result<bool, ValidationError>

pub fn append(&mut self, node: Vec<u8>) -> Result<(), ValidationError>