Module indy_vdr::ledger

Request and response types for the Indy Node ledger

Modules

constants
Ledger transaction type identifiers

identifiers
Identifiers for stored objects on the ledger

requests
Types for constructing ledger transactions

Structs

PreparedRequest
A ledger transaction request which has been prepared for dispatch

RequestBuilder
A utility class for constructing ledger transaction requests