Trait indy_vdr::ledger::requests::RequestType

pub trait RequestType: Serialize {
    fn get_txn_type<'a>() -> &'a str;

    fn get_sp_key(
        &self, 
        _protocol_version: ProtocolVersion
    ) -> VdrResult<Option<Vec<u8>>> { ... }

    fn get_sp_timestamps(&self) -> VdrResult<(Option<u64>, Option<u64>)> { ... }
}

Base trait for all ledger transaction request operations

Required methods

fn get_txn_type<'a>() -> &'a str

Get the transaction type as a numeric string

Loading content...
Provided methods

fn get_sp_key(
&self,
_protocol_version: ProtocolVersion
) -> VdrResult<Option<Vec<u8>>>

Get a state proof key for the transaction, if any can be derived

fn get_sp_timestamps(&self) -> VdrResult<(Option<u64>, Option<u64>)>

Get the state proof timestamps for the request, if any

Loading content...