Struct indy_vdr::ledger::RequestBuilder

pub struct RequestBuilder {
    pub protocol_version: ProtocolVersion,
}

A utility class for constructing ledger transaction requests

Fields
protocol_version: ProtocolVersion
Methods

impl RequestBuilder

pub fn new(protocol_version: ProtocolVersion) -> Self

Create a new RequestBuilder for a specific protocol version

pub fn build<T: RequestType>(
&self,
operation: T,
identifier: Option<&DidValue>
) -> VdrResult<PreparedRequest>

Build a generic prepared request

pub fn build_nym_request(
&self,
identifier: &DidValue,
dest: &DidValue,
verkey: Option<String>,
alias: Option<String>,
role: Option<String>
) -> VdrResult<PreparedRequest>

Build a NYM transaction request

pub fn build_get_nym_request(
&self,
identifier: Option<&DidValue>,
dest: &DidValue
) -> VdrResult<PreparedRequest>

Build a GET_NYM transaction request

pub fn build_attrib_request(
&self,
identifier: &DidValue,
dest: &DidValue,
hash: Option<String>,
raw: Option<&SJsonValue>,
enc: Option<String>
) -> VdrResult<PreparedRequest>

Build an ATTRIB transaction request

pub fn build_get_attrib_request(
&self,
identifier: Option<&DidValue>,
dest: &DidValue,
raw: Option<String>,
hash: Option<String>,
enc: Option<String>
) -> VdrResult<PreparedRequest>

Build a GET_ATTRIB transaction request

pub fn build_node_request(
&self,
identifier: &DidValue,
dest: &DidValue,
data: NodeOperationData
) -> VdrResult<PreparedRequest>

Build a NODE transaction request

pub fn build_get_validator_info_request(
&self,
identifier: &DidValue
) -> VdrResult<PreparedRequest>

Build a GET_VALIDATOR_INFO transaction request

pub fn build_get_txn_request(
&self,
identifier: Option<&DidValue>,
ledger_type: i32,
seq_no: i32
) -> VdrResult<PreparedRequest>

Build a GET_TXN transaction request

pub fn build_pool_config(
&self,
identifier: &DidValue,
writes: bool,
force: bool
) -> VdrResult<PreparedRequest>

Build a POOL_CONFIG transaction request

pub fn build_pool_restart(
&self,
identifier: &DidValue,
action: &str,
datetime: Option<&str>
) -> VdrResult<PreparedRequest>

Build a POOL_RESTART transaction request

pub fn build_pool_upgrade(
&self,
identifier: &DidValue,
name: &str,
version: &str,
action: &str,
sha256: &str,
timeout: Option<u32>,
schedule: Option<Schedule>,
justification: Option<&str>,
reinstall: bool,
force: bool,
package: Option<&str>
) -> VdrResult<PreparedRequest>

Build a POOL_UPGRADE transaction request

pub fn build_auth_rule_request(
&self,
submitter_did: &DidValue,
txn_type: String,
action: String,
field: String,
old_value: Option<String>,
new_value: Option<String>,
constraint: Constraint
) -> VdrResult<PreparedRequest>

Build an AUTH_RULE transaction request

pub fn build_auth_rules_request(
&self,
submitter_did: &DidValue,
rules: AuthRules
) -> VdrResult<PreparedRequest>

Build an AUTH_RULES transaction request

pub fn build_get_auth_rule_request(
&self,
submitter_did: Option<&DidValue>,
auth_type: Option<String>,
auth_action: Option<String>,
field: Option<String>,
old_value: Option<String>,
new_value: Option<String>
) -> VdrResult<PreparedRequest>

Build a GET_AUTH_RULE transaction request

pub fn build_txn_author_agreement_request(
&self,
identifier: &DidValue,
text: Option<String>,
version: String,
ratification_ts: Option<u64>,
retirement_ts: Option<u64>
) -> VdrResult<PreparedRequest>

Build a TXN_AUTHR_AGRMT transacation request

pub fn build_get_txn_author_agreement_request(
&self,
identifier: Option<&DidValue>,
data: Option<&GetTxnAuthorAgreementData>
) -> VdrResult<PreparedRequest>

Build a GET_TXN_AUTHR_AGRMT transaction request

pub fn build_disable_all_txn_author_agreements_request(
&self,
identifier: &DidValue
) -> VdrResult<PreparedRequest>

Build a DISABLE_ALL_TXN_AUTHR_AGRMTS transaction request

pub fn build_acceptance_mechanisms_request(
&self,
identifier: &DidValue,
aml: AcceptanceMechanisms,
version: String,
aml_context: Option<String>
) -> VdrResult<PreparedRequest>

Build a TXN_AUTHR_AGRMT_AML transaction request

pub fn build_get_acceptance_mechanisms_request(
&self,
identifier: Option<&DidValue>,
timestamp: Option<u64>,
version: Option<String>
) -> VdrResult<PreparedRequest>

Build a GET_TXN_AUTHR_AGRMT_AML transaction request

pub fn build_schema_request(
&self,
identifier: &DidValue,
schema: Schema
) -> VdrResult<PreparedRequest>

Build a SCHEMA transaction request

pub fn build_get_schema_request(
&self,
identifier: Option<&DidValue>,
id: &SchemaId
) -> VdrResult<PreparedRequest>

Build a GET_SCHEMA transaction request

pub fn build_cred_def_request(
&self,
identifier: &DidValue,
cred_def: CredentialDefinition
) -> VdrResult<PreparedRequest>

Build a CRED_DEF transaction request

pub fn build_get_cred_def_request(
&self,
identifier: Option<&DidValue>,
id: &CredentialDefinitionId
) -> VdrResult<PreparedRequest>

Build a GET_CRED_DEF transaction request

pub fn build_get_revoc_reg_def_request(
&self,
identifier: Option<&DidValue>,
id: &RevocationRegistryId
) -> VdrResult<PreparedRequest>

Build a GET_REVOC_REG_DEF transaction request

pub fn build_get_revoc_reg_request(
&self,
identifier: Option<&DidValue>,
revoc_reg_def_id: &RevocationRegistryId,
timestamp: i64
) -> VdrResult<PreparedRequest>

Build a GET_REVOC_REG transaction request

pub fn build_get_revoc_reg_delta_request(
&self,
identifier: Option<&DidValue>,
revoc_reg_def_id: &RevocationRegistryId,
from: Option<i64>,
to: i64
) -> VdrResult<PreparedRequest>

Build a GET_REVOC_REG_DELTA transaction request

pub fn build_revoc_reg_def_request(
&self,
identifier: &DidValue,
revoc_reg: RevocationRegistryDefinition
) -> VdrResult<PreparedRequest>

Build a REVOC_REG_DEF transaction request

pub fn build_revoc_reg_entry_request(
&self,
identifier: &DidValue,
revoc_reg_def_id: &RevocationRegistryId,
revoc_def_type: &RegistryType,
rev_reg_entry: RevocationRegistryDelta
) -> VdrResult<PreparedRequest>

Build a REVOC_REG_ENTRY transaction request

pub fn prepare_txn_author_agreement_acceptance_data(
&self,
text: Option<&str>,
version: Option<&str>,
hash: Option<&str>,
mechanism: &str,
time: u64
) -> VdrResult<TxnAuthrAgrmtAcceptanceData>

Prepare transaction author agreement acceptance data

pub fn build_custom_request_from_slice(
&self,
message: &[u8],
sp_key: Option<Vec<u8>>,
sp_timestamps: (Option<u64>, Option<u64>)
) -> VdrResult<(PreparedRequest, Option<RequestTarget>)>

Build a custom ledger transaction request from a byte array

pub fn build_custom_request_from_str(
&self,
message: &str,
sp_key: Option<Vec<u8>>,
sp_timestamps: (Option<u64>, Option<u64>)
) -> VdrResult<(PreparedRequest, Option<RequestTarget>)>

Build a custom ledger transaction request from a string

pub fn build_custom_request(
&self,
req_json: SJsonValue,
sp_key: Option<Vec<u8>>,
sp_timestamps: (Option<u64>, Option<u64>)
) -> VdrResult<(PreparedRequest, Option<RequestTarget>)>

Build a custom ledger transaction request from a JSON value

pub fn build_rich_schema_request(
&self,
identifier: &DidValue,
rs_schema: RichSchema
) -> VdrResult<PreparedRequest>

Build a RICH_SCHEMA transaction request

pub fn build_get_rich_schema_by_id(
&self,
identifier: &DidValue,
rs_id: &RichSchemaId
) -> VdrResult<PreparedRequest>

Build a GET_RICH_SCHEMA_BY_ID transaction request

pub fn build_get_rich_schema_by_metadata(
&self,
identifier: &DidValue,
rs_type: i32,
rs_name: String,
rs_version: String
) -> VdrResult<PreparedRequest>

Build a GET_RICH_SCHEMA_BY_METADATA transaction request