Module indy_vdr::pool

Handling of verifier pool instances and communication

Modules

helpers
Methods for performing requests against the verifier pool

networker
Pool networker traits and implementations

requests
Data types and traits for handling pending verifier pool requests

Structs

PoolBuilder
A utility class for building a new pool instance or runner.

PoolImpl
The default Pool implementation

PoolRunner
The PoolRunner instance creates a separate thread for handling pool events, allowing the use of callbacks instead of async functions for interacting with the pool as well as simplifying validator pool refreshes.

PoolRunnerStatus
The current status of a validator pool.

PoolSetup
Setup parameters for PoolImpl, shared with cloned instances and derived pool requests

PoolTransactions
A collection of pool genesis transactions.

RequestHandle

VerifierInfo
Validator node details loaded from pool transactions

VerifierKey
A verifier node BLS signing key

Enums

LedgerType
An Indy Node subledger identifier

ProtocolVersion
The Indy Node communication protocol version

RequestResult
Common result type returned by request handlers

RequestTarget
Enum distinguishing between different ledger request handling methods

SingleReply
The result of a request to a single validator node

Traits

Pool
A generic verifier pool with support for creating pool transaction requests

Type Definitions

LocalPool
A verifier pool instance restricted to a single thread

NodeReplies
A collection of results from multiple validator nodes

SharedPool
A verifier pool instance which can be shared between threads

TimingResult
Type representing timing information collected for ledger transaction request

VerifierKeys
A collection of verifier node BLS keys

Verifiers
A standard collection of verifier information, indexed by node alias