Module indy_vdr::pool::networker

Pool networker traits and implementations

Structs

MakeLocal
A derived NetworkerFactory producing thread-local cloneable instances

MakeShared
A derived NetworkerFactory producing shareable instances

ZMQNetworker
ZeroMQ Networker implementation

ZMQNetworkerFactory
ZeroMQ NetworkerFactory implementation

Enums

NetworkerEvent
Events used to drive a Networker instance

Traits

Networker
A simple trait implemented by all networker types

NetworkerFactory
A factory for Networker instances

Type Definitions

LocalNetworker
A Networker instance which can be cloned and used within one thread

SharedNetworker
A Networker instance which can be cloned and used across multiple threads