Struct indy_vdr::pool::PoolTransactions

pub struct PoolTransactions { /* fields omitted */ }

A collection of pool genesis transactions.

Methods

impl PoolTransactions

pub fn from_json(txns: &str) -> VdrResult<Self>

Load JSON pool transactions from a string.

pub fn from_json_file<P>(file_name: P) -> VdrResult<Self> where
P: AsRef<Path> + Debug,

Load JSON pool transactions from a file.

pub fn from_json_transactions<T>(txns: T) -> VdrResult<Self> where
T: IntoIterator,
T::Item: AsRef<str>,

Load pool transactions from a sequence of JSON strings.

pub fn from_transactions<T>(txns: T) -> Self where
T: IntoIterator,
T::Item: AsRef<[u8]>,

Load pool transactions from a sequence of msgpack-encoded byte strings.

pub fn extend<T>(&mut self, txns: T) where
T: IntoIterator<Item = Vec<u8>>,

Extend the pool transactions with a set of msgpack-encoded byte strings.

pub fn extend_from_json<'a, T>(&mut self, txns: T) -> VdrResult<()> where
T: IntoIterator,
T::Item: AsRef<str>,

Extend the pool transactions with a set of JSON strings.

pub fn merkle_tree(&self) -> VdrResult<MerkleTree>

Derive a MerkleTree instance from the set of pool transactions.

pub fn into_merkle_tree(self) -> VdrResult<MerkleTree>

Convert the set of pool transactions into a MerkleTree instance.

pub fn iter(&self) -> impl Iterator<Item = &Vec<u8>>

Iterate the set of transactions as a sequence of msgpack-encoded byte strings.

pub fn encode_json(&self) -> VdrResult<Vec<String>>

Get a sequence of JSON strings representing the pool transactions.

pub fn json_values(&self) -> VdrResult<Vec<SJsonValue>>

Get a sequence of serde_json::Value instances representing the pool transactions.

pub fn len(&self) -> usize

Get the number of pool transactions.