Trait indy_vdr::pool::Pool

pub trait Pool: Clone {
    type Request: PoolRequest;
    fn get_config(&self) -> &PoolConfig;

    fn create_request<'a>(
        &'a self, 
        req_id: String, 
        req_json: String
    ) -> LocalBoxFuture<'a, VdrResult<Self::Request>>;

    fn get_merkle_tree(&self) -> &MerkleTree;

    fn get_node_aliases(&self) -> Vec<String>;

    fn get_merkle_tree_info(&self) -> (String, usize) { ... }

    fn get_request_builder(&self) -> RequestBuilder { ... }

    fn get_json_transactions(&self) -> VdrResult<Vec<String>> { ... }
}

A generic verifier pool with support for creating pool transaction requests

Associated Types

type Request: PoolRequest

Loading content...
Required methods

fn get_config(&self) -> &PoolConfig

Get the pool configuration for this instance

fn create_request<'a>(
&'a self,
req_id: String,
req_json: String
) -> LocalBoxFuture<'a, VdrResult<Self::Request>>

Create a new pool request instance

fn get_merkle_tree(&self) -> &MerkleTree

Get the merkle tree representing the verifier pool transactions

fn get_node_aliases(&self) -> Vec<String>

A sequence of verifier node aliases

Loading content...
Provided methods

fn get_merkle_tree_info(&self) -> (String, usize)

Get the size and root of the pool transactions merkle tree

fn get_request_builder(&self) -> RequestBuilder

Get a request builder corresponding to this verifier pool

fn get_json_transactions(&self) -> VdrResult<Vec<String>>

Get the set of verifier pool transactions in JSON format

Loading content...